Move Task15.cpp
