+++ Sem. 02/Pract.02/Task15.cpp
@@ -1,0 +1,21 @@
+// Да се състави програма, чрез която се въвежда 4-цифренo естествено число от интервала [1000.. 9999]. 
+// От това число се формират 2 нови 2-цифрени числа. Първото число се формира от 1-та и 4-та цифра 
+// на въведеното число. Второто число се формира от 2-рa - 3-та цифра на въведеното число. 
+// На екрана да се изведе дали 1-то ново число e по-малко <, равно = или по-голямо от 2-то число.
+#include <iostream>;
+int main() {
+    int number;
+    std::cin >> number;
+    int num1 = (number / 1000) * 10 + number % 10;
+    int num2 = (number / 100 % 10) * 10 + (number / 10) % 10;
+
+    if (num1 < num2) {
+        std::cout << "less ( " << num1 << " < " << num2 << " )\n";
+    }
+    else if (num1 > num2) {
+        std::cout << "greater ( " << num1 << " > " << num2 << " )\n";
+    }
+    else {
+        std::cout << "equal ( " << num1 << " = " << num2 << " )\n";
+    }
+}
